Reduce despawn code duplication
