+++ Assets/Prefabs/Enemies/Jackolatern.prefab
@@ -16,6 +16,7 @@ GameObject:
   - component: {fileID: 4670997850951537308}
   - component: {fileID: 4670997850951537314}
   - component: {fileID: -8866487689074243921}
+  - component: {fileID: -3618504319965289538}
   m_Layer: 6
   m_Name: Jackolatern
   m_TagString: Enemy
@@ -105,8 +106,6 @@ MonoBehaviour:
   IgnoreMe:
     serializedVersion: 2
     m_Bits: 192
-  aimRate: 1.5
-  fireRate: 1.5
   range: 20
   fireForceMinimum: 200
   fireForceMaximum: 350
@@ -193,5 +192,17 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 04fe750e87560bc41ab2b029d4812bbb, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &-3618504319965289538
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4670997851035706755}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e4ef37c6ecdb6cc41965768b3e92d1ca, type: 3}
   m_Name: 
   m_EditorClassIdentifier:
