+++ Assets/Prefabs/Graves/GraveObstacleSmall.prefab
@@ -12,7 +12,8 @@ GameObject:
   - component: {fileID: 1349739426613462087}
   - component: {fileID: 1349739426613462084}
   - component: {fileID: 1349739426613462085}
-  - component: {fileID: 1349739426613462139}
+  - component: {fileID: 825258746188476305}
+  - component: {fileID: -8051218455254796603}
   m_Layer: 7
   m_Name: GraveObstacleSmall
   m_TagString: Obstacle
@@ -134,18 +135,6 @@ Rigidbody2D:
   m_SleepingMode: 1
   m_CollisionDetection: 0
   m_Constraints: 4
---- !u!114 &1349739426613462139
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1349739426613462080}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 39ecf63b7a9390a4bad07f86af84e960, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!114 &825258746188476305
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -156,5 +145,17 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 34a0c90d7f351df46a17b5f746488068, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &-8051218455254796603
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1349739426613462080}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e4ef37c6ecdb6cc41965768b3e92d1ca, type: 3}
   m_Name: 
   m_EditorClassIdentifier:
