--- Assets/Scripts/Spawnables/Obstacles/DespawnObstacle.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 39ecf63b7a9390a4bad07f86af84e960
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
