+++ Assets/Scripts/Spawnables/Shared.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9e0ea4630a1bcf446a148156e491e642
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
