+++ Assets/Scripts/Spawnables/Shared/DespawnOnOutOfBounds.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1bb3517cfb5e2ea4f8397047bad0557c
+guid: e4ef37c6ecdb6cc41965768b3e92d1ca
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
